% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTADs.R
\name{call_hTADs}
\alias{call_hTADs}
\title{Call hierarchical TADs}
\usage{
call_hTADs(file_name, cores = 1, max_pcs = 200, method = c("fast",
  "accurate"), n_samples = 60, plot = FALSE)
}
\arguments{
\item{file_name}{The path to the file to read.}

\item{cores}{When `method` is `"accurate"`, the number of cores to use for parallel execution.}

\item{max_pcs}{The maximum number of principal components to retain for the analysis.}

\item{method}{Which version of the algorithm to use.}

\item{n_samples}{When `method` is `"fast"`, the number of samples used to approximate the optimal solution.}

\item{plot}{Logical. Whether to plot the scores of every tested `n_pcs`/`n_clusters` combination.}
}
\value{
`htad` object that defines the clustering of genomic regions.
}
\description{
Computes a constrained hierarchical clustering of genomic regions in a HiC experiment,
choosing the optimal amount of information from the HiC matrix and selecting the most informative number of TADs.
}
\examples{
htads <- call_hTADs("file_name.abc")
}
