% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TADpole.R
\name{CH_map}
\alias{CH_map}
\title{Plot matrix of Calinski-Harabasz (CH) index}
\usage{
CH_map(tadpole)
}
\arguments{
\item{TADpole}{`TADpole` object returned as by function `TADpole`.}
}
\description{
Plot matrix of Calinski-Harabasz (CH) index
}
\examples{
mat_file <- system.file("extdata", "raw_chr18:460-606_20kb.tsv", package = "TADpole")
tadpole <- TADpole(mat_file)
CH_map(tadpole)
}
