% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffT.R
\name{diffT}
\alias{diffT}
\title{Compute diffT score between two TAD calls}
\usage{
diffT(bed_x, bed_y)
}
\arguments{
\item{`bed_x, }{bed_y` two `data.frame`s with a BED-like format with 3 columns: chromosome, start and end coordinates of each TAD, in bins.}
}
\description{
Compute diffT score between two TAD calls
}
\examples{
difft_control_case <- diffT(control, case)
}
