% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TADpole.R
\name{load_mat}
\alias{load_mat}
\title{Load a Hi-C matrix from a file}
\usage{
load_mat(
  mat_file,
  chr,
  start,
  end,
  resol,
  bad_frac = 0.01,
  centromere_search = FALSE
)
}
\arguments{
\item{mat_file}{path to the input file. Must be in a tab-delimited matrix format.}

\item{chr}{string with the chromosome name.}

\item{start}{numeric start position of the region or of the chromosome.}

\item{end}{numeric end position of the region or of the chromosome.}

\item{resol}{numeric resolution/binning of the Hi-C experiment.}

\item{bad_frac}{fraction of the matrix to falg as bad rows/columns.}

\item{centromere_search}{split the matrix by the centrormere into two, smaller matrices representing the chromosomal arms. Useful when working with big (>15000 bins) datasets.}
}
\description{
Load a Hi-C matrix from a file
}
\examples{
mat_file <- system.file("extdata", "raw_chr18:460-606_20kb.tsv", package = "TADpole")
mat <- load_mat(mat_file, chr = "chr18", start = 496, end = 606, resol = 20000)
}
