% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TADpole.R
\name{plot_borders}
\alias{plot_borders}
\title{Plot borders}
\usage{
plot_borders(tadpole, mat_file, centromere_search)
}
\arguments{
\item{mat_file}{path to the input file. Must be in a tab-delimited matrix format.}

\item{centromere_search}{split the matrix by the centrormere into two, smaller matrices representing the chromosomal arms. Useful when working with big (>15000 bins) datasets.}

\item{TADpole}{`TADpole` object returned as by function `TADpole`.}
}
\description{
Plot borders
}
\examples{
tadpole <- TADpole('data/chromosome18_10Mb.tsv')
plot_borders(tadpole, input_data = 'data/chromosome18_10Mb.tsv',centromere_search = FALSE)
}
