% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTADs.R
\name{plot_dendro}
\alias{plot_dendro}
\title{Plot dendrogram}
\usage{
plot_dendro(htads)
}
\arguments{
\item{htads}{`htad` object returned by `call_HTADs`.}
}
\description{
Plot dendrogram
}
\examples{
load('data/chromosome18_10Mb.Rdata')
htads <- call_HTADs(chromosome18_10Mb)
plot_dendro(htads)
}
