% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TADpole.R
\name{plot_hierarchy}
\alias{plot_hierarchy}
\title{Plot hierarchical_plot}
\usage{
plot_hierarchy(
  mat_file,
  tadpole,
  chr,
  start,
  end,
  resol,
  centromere_search = FALSE
)
}
\arguments{
\item{mat_file}{path to the input file. Must be in a tab-delimited matrix format.}

\item{chr}{string with the chromosome name.}

\item{start}{numeric start position of the region or of the chromosome.}

\item{end}{numeric end position of the region or of the chromosome.}

\item{resol}{numeric resolution/binning of the Hi-C experiment.}

\item{centromere_search}{split the matrix by the centrormere into two, smaller matrices representing the chromosomal arms. Useful when working with big (>15000 bins) datasets.}

\item{TADpole}{`TADpole` object returned as by function `TADpole`.}
}
\description{
Plot hierarchical_plot
}
\examples{
mat_file <- system.file("extdata", "raw_chr18:460-606_20kb.tsv", package = "TADpole")
tadpole <- TADpole(mat_file, chr = "chr18", start = 496, end = 606, resol = 20000)
plot_hierarchy(mat_file, tadpole, chr = "chr18", start = 496, end = 606, resol = 20000)
}
