% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTADs.R
\name{plot_scores}
\alias{plot_scores}
\title{Plot the Caliski-Harabasz index of every tested `n_pcs`/`n_clusters` combination}
\usage{
plot_scores(htads)
}
\arguments{
\item{htads}{`htad` object returned by `call_HTADs`.}
}
\description{
Plot the Caliski-Harabasz index of every tested `n_pcs`/`n_clusters` combination
}
\examples{
plot_scores(htads)
}
