% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTADs.R
\name{plot_var}
\alias{plot_var}
\title{Plot retained variance}
\usage{
plot_var(input_data, max_pcs = NULL, mark = 200)
}
\arguments{
\item{input_data}{`data.frame` with 3 columns containing HiC data in the format `(bin1, bin2, score)`.}

\item{max_pcs}{The maximum number of principal components whose cumulative variance is to be plotted.}

\item{mark}{A place on the abcissa where to put a visual mark.}
}
\description{
Plot retained variance
}
\examples{
load('data/chromosome18_10Mb.Rdata')
plot_var(chromosome18_10Mb)
}
