% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffT.R
\name{random_bed}
\alias{random_bed}
\title{Compute a random set of TAD border coordinates from a sample partition}
\usage{
random_bed(bed, bad_columns = NULL)
}
\arguments{
\item{`bed`}{a `data.frame` with a BED-like format with 3 columns: chromosome, start and end coordinates of each TAD, in bins.}

\item{`bad_columns`}{a numeric `vector` with the positions of the bad columns. TAD borders will not be placed on bad columns. Default value of `NULL` means no bad columns will be introduced.}
}
\description{
Compute a random set of TAD border coordinates from a sample partition
}
\examples{
control <- read.table(system.file("extdata", "control.bed", package = "TADpole"))
random_coords <- random_bed(control)
}
