% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTADs.R
\name{save_bed_coordinates}
\alias{save_bed_coordinates}
\title{Save Bed file with the TAD's borders coordinates.}
\usage{
save_bed_coordinates(htads, chrom, file = "")
}
\arguments{
\item{htads}{`htad` object returned by `call_HTADs`.}

\item{chrom}{`character`. The name of the chromosome.}

\item{file}{either a character string naming a file or a `connection` open for writing. `""` indicates output to the console.}
}
\description{
Save Bed file with the TAD's borders coordinates.
}
\examples{
load('data/chromosome18_10Mb.Rdata')
htads <- call_HTADs(chromosome18_10Mb)
save_bed_coordinates(htads, chrom = 'chr18', file = 'chr18_borders.bed')
}
