% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TADpole.R
\name{TADpole}
\alias{TADpole}
\title{Call hierarchical TADs}
\usage{
TADpole(mat_file, max_pcs = 200, min_clusters = 2, bad_frac = 0.01,
  centromere_search = FALSE, hist_bad_columns = FALSE)
}
\arguments{
\item{mat_file}{path to the input file. Must be in a tab-delimited matrix format.}

\item{max_pcs}{The maximum number of principal components to retain for the analysis.}

\item{min_clusters}{Minimum number of clusters into which partition the chromosome.}

\item{bad_frac}{fraction of the matrix to falg as bad rows/columns.}

\item{centromere_search}{split the matrix by the centrormere into two smaller matrices representing the chromosomal arms. Useful when woring with big (>15000 bins) datasets.}

\item{hist_bad_columns}{plot the distribution of row/column coverage to help in selecting a useful value for `bad_frac`. Mostly for debugging.}
}
\value{
`tadpole` object that defines the clustering of genomic regions.
}
\description{
Computes a constrained hierarchical clustering of genomic regions in a HiC experiment,
choosing the optimal amount of information from the HiC matrix and selecting the most informative number of TADs.
}
\details{
The `centromere_search` parameter will split the matrix into two by the region with the longes stretch of bad (low signal) rows/columns.
It will do so regardless of whether this stretch represents a true centromere or not. Note that this feature is useful when processing an entire chromosome,
but be cautious of interpreting the partitions as the two chromosomal arms (p and q) when working with smaller regions.
}
\examples{
chromosome18_10Mb <- system.file("extdata", "chromosome18_10Mb.tsv", package = "TADpole")
tadpole <- TADpole(chromosome18_10Mb)
}
