% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTADs.R
\name{call_HTADs}
\alias{call_HTADs}
\title{Call hierarchical TADs}
\usage{
call_HTADs(input_data, cores = 1, max_pcs = 200,
  method = c("accurate", "fast"), n_samples = 60, min_clusters = 1)
}
\arguments{
\item{input_data}{`data.frame` with 3 columns containing HiC data in the format `(bin1, bin2, score)`.}

\item{cores}{When `method` is `"accurate"`, the number of cores to use for parallel execution.}

\item{max_pcs}{The maximum number of principal components to retain for the analysis.}

\item{method}{Which version of the algorithm to use.}

\item{n_samples}{When `method` is `"fast"`, the number of samples used to approximate the optimal solution.}

\item{min_clusters}{Minimum number of clusters into which partition the chromosome.}

\item{plot}{Logical. Whether to plot the scores of every tested `n_pcs`/`n_clusters` combination.}
}
\value{
`htad` object that defines the clustering of genomic regions.
}
\description{
Computes a constrained hierarchical clustering of genomic regions in a HiC experiment,
choosing the optimal amount of information from the HiC matrix and selecting the most informative number of TADs.
}
\examples{
load('data/chromosome18_10Mb.Rdata')
htads <- call_HTADs(chromosome18_10Mb)
}
