% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTADs.R
\name{load_mat}
\alias{load_mat}
\title{Convert a 3-column data frame to a symmetric HiC matrix}
\usage{
load_mat(input_data)
}
\arguments{
\item{input_data}{`data.frame` with 3 columns containing HiC data in the format `(bin1, bin2, score)`.}
}
\description{
Convert a 3-column data frame to a symmetric HiC matrix
}
