% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TADpole.R
\name{load_mat}
\alias{load_mat}
\title{Load a Hi-C matrix from a file}
\usage{
load_mat(mat_file, bad_frac = 0.01, centromere_search = FALSE,
  hist_bad_columns = FALSE)
}
\arguments{
\item{mat_file}{path to the input file. Must be in a tab-delimited matrix format.}

\item{bad_frac}{fraction of the matrix to falg as bad rows/columns.}

\item{centromere_search}{split the matrix by the centrormere into two, smaller matrices representing the chromosomal arms. Useful when working with big (>15000 bins) datasets.}

\item{hist_bad_columns}{plot the distribution of row/column coverage to help in selecting a useful value for `bad_frac`. Mostly for debugging.}
}
\description{
Load a Hi-C matrix from a file
}
\examples{
chromosome18_10Mb <- system.file("extdata", "chromosome18_10Mb.tsv", package = "TADpole")
mat <- load_mat(chromosome18_10Mb)
}
