% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TADpole.R
\name{plot_borders}
\alias{plot_borders}
\title{Plot borders}
\usage{
plot_borders(tadpole, mat_file, centromere_search = FALSE)
}
\arguments{
\item{mat_file}{path to the input file. Must be in a tab-delimited matrix format.}

\item{centromere_search}{split the matrix by the centrormere into two, smaller matrices representing the chromosomal arms. Useful when working with big (>15000 bins) datasets.}

\item{TADpole}{`TADpole` object returned as by function `TADpole`.}
}
\description{
Plot borders
}
\examples{
chromosome18_10Mb <- system.file("extdata", "chromosome18_10Mb.tsv", package = "TADpole")
tadpole <- TADpole(chromosome18_10Mb)
plot_borders(tadpole, chromosome18_10Mb)
}
