% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTADs.R
\name{plot_borders}
\alias{plot_borders}
\title{Plot TAD borders on a HiC matrix}
\usage{
plot_borders(htads, input_data)
}
\arguments{
\item{htads}{`htad` object returned by `call_HTADs`.}

\item{input_data}{`data.frame` with 3 columns containing HiC data in the format `(bin1, bin2, score)`.}

\item{add.dendro}{Logical. If `TRUE`, plot the clustering dendrogram beside the HiC matrix.}
}
\description{
Plot TAD borders on a HiC matrix
}
\examples{
load('data/chromosome18_10Mb.Rdata')
htads <- call_HTADs(chromosome18_10Mb)
plot_borders(htads, chromosome18_10Mb)
}
