% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTADs.R
\name{plot_borders}
\alias{plot_borders}
\title{TAD border detection plot}
\usage{
plot_borders(matrix, htads_output)
}
\arguments{
\item{mat}{`mat` object returned by `load_mat`.}

\item{htads}{`htad` object returned by `call_HTADs`.}
}
\description{
TAD border detection plot
}
\examples{
load('data/chromosome18_10Mb.Rdata')
htads <- call_hTADs(chromosome18_10Mb)
plot_borders(mat,htads)
}
