% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TADpole.R
\name{plot_dendro}
\alias{plot_dendro}
\title{Plot dendrogram}
\usage{
plot_dendro(tadpole)
}
\arguments{
\item{TADpole}{`TADpole` object returned as by function `TADpole`.}
}
\description{
Plot dendrogram
}
\examples{
chromosome18_10Mb <- system.file("extdata", "chromosome18_10Mb.tsv", package = "TADpole")
tadpole <- TADpole(chromosome18_10Mb)
plot_dendro(tadpole)
}
